示波器v3节点项目
基于 PyQt5 + nodeeditor 的信号路由图编辑器扩展，新增示波器 V3 节点。

核心功能：
- 双线程架构（数据采集 + 渲染分离）
- 环形缓冲区 RingBuffer 存储波形数据
- 双缓冲 QPainterPath 波形绘制
- 网格 / 刻度 / 多通道支持
- 心跳检测 + 背压保护
- 鼠标交互（滚轮缩放、拖拽偏移/滚动）
- 刻度绘制（水平时间 + 垂直电压标注）
- 数据导出 JSON

+++ connnodes/waveform_protocol_v3.py
@@ -86,7 +86,7 @@ def decode_raw(buf_bytes: bytes, pos: int, data_count: int):
     data = np.frombuffer(
         buf_bytes[pos + _DATA_PAYLOAD_OFFSET:pos + _DATA_PAYLOAD_OFFSET + 4 * data_count],
         dtype=np.float32
-    ).astype(np.float64)
+    )
     return channel_id, data, gap_count, interval_us
 
 
@@ -127,8 +127,8 @@ def decode_packet(ba: QByteArray) -> Optional[dict]:
     if not _check_crc(ba, crc_offset):
         return None
 
-    # 6. 解码数据（numpy float64）
-    data = np.frombuffer(raw[13:13 + 4 * data_count], dtype=np.float32).astype(np.float64)
+    # 6. 解码数据（numpy float32）
+    data = np.frombuffer(raw[13:13 + 4 * data_count], dtype=np.float32)
 
     return {
         "channel_id": channel_id,
